Combine assets/assetManifest modules
