+++ src/components/Form/Fields/File.js
@@ -6,7 +6,7 @@ import Dropzone from 'react-dropzone';
 import { uniqueId } from 'lodash';
 import cx from 'classnames';
 import { fieldPropTypes } from 'redux-form';
-import { actionCreators as assetActions } from '../../../ducks/modules/protocol/assets';
+import { actionCreators as assetActions } from '../../../ducks/modules/protocol/assetManifest';
 
 class FileInput extends PureComponent {
   static propTypes = {
