--- src/ducks/modules/protocol/__tests__/externalData.test.js
@@ -1,16 +1,0 @@
-/* eslint-env jest */
-
-import reducer from '../externalData';
-
-describe('protocol.externalData', () => {
-  it('does nothing', () => {
-    const noop = reducer(
-      {},
-      {},
-    );
-
-    expect(
-      noop,
-    ).toEqual({});
-  });
-});
